Moved Casual CSS to its own function
